database: Add watchtower module.
